fixed all float/subclass issues with deserialization of the database. The insert and query functions should now wok.

+++ molSimplifyAD/post_classes.py
@@ -212,21 +212,21 @@ class DFTRun(object):
             
     def obtain_ML_dists(self):
         try:
-            self.mind = minimum_ML_dist(self.mol)
-            self.maxd = maximum_any_dist(self.mol)
-            self.meand = mean_ML_dist(self.mol)
+            self.mind = float(minimum_ML_dist(self.mol))
+            self.maxd = float(maximum_any_dist(self.mol))
+            self.meand = float(mean_ML_dist(self.mol))
             ax_dist, eq_dist = getOctBondDistances(self.mol)
             if len(ax_dist) < 2:
                 ax_dist.append(ax_dist[0])
-            self.ax1_MLB = np.mean(ax_dist[0])
-            self.ax2_MLB = np.mean(ax_dist[1])
+            self.ax1_MLB = float(np.mean(ax_dist[0]))
+            self.ax2_MLB = float(np.mean(ax_dist[1]))
             total_eq_distance = 0
             counter = 0
             for eqligs in eq_dist:
                 for bonds in eqligs:
                     total_eq_distance += bonds
                     counter += 1
-            self.eq_MLB = total_eq_distance / counter
+            self.eq_MLB = float(total_eq_distance / counter)
         except:
             # self.coord = 'error'
             self.eq_MLB = 'error'
@@ -237,15 +237,15 @@ class DFTRun(object):
             ax_dist, eq_dist = getOctBondDistances(self.init_mol)
             if len(ax_dist) < 2:
                 ax_dist.append(ax_dist[0])
-            self.init_ax1_MLB = np.mean(ax_dist[0])
-            self.init_ax2_MLB = np.mean(ax_dist[1])
+            self.init_ax1_MLB = float(np.mean(ax_dist[0]))
+            self.init_ax2_MLB = float(np.mean(ax_dist[1]))
             total_eq_distance = 0
             counter = 0
             for eqligs in eq_dist:
                 for bonds in eqligs:
                     total_eq_distance += bonds
                     counter += 1
-            self.init_eq_MLB = total_eq_distance / counter
+            self.init_eq_MLB = float(total_eq_distance / counter)
         except:
             # self.init_coord = 'error'
             self.init_eq_MLB = 'error'
